Extract funding values functionality
