+++ lib/pafs_core.rb
@@ -10,6 +10,7 @@ require "pafs_core/urgency"
 require "pafs_core/errors"
 require "pafs_core/form_builder"
 require "pafs_core/funding_sources"
+require "pafs_core/funding_values"
 require "pafs_core/risks"
 require "pafs_core/standard_of_protection"
 require "pafs_core/environmental_outcomes"
